chore: refactor to reader.py and writer.py
